backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==1.4.46
python-dotenv==1.0.0
twilio==8.13.0
python-multipart==0.0.6